KYC refactoring
