+++ src/admin/admin.controller.ts
@@ -15,14 +15,14 @@ import { ApiBearerAuth, ApiExcludeEndpoint } from '@nestjs/swagger';
 import { RoleGuard } from 'src/shared/auth/role.guard';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { MailService } from 'src/shared/services/mail.service';
-import { IdentService } from 'src/user/models/ident/ident.service';
+import { KycService } from 'src/user/models/kyc/kyc.service';
 import { getConnection } from 'typeorm';
 import { SendMailDto } from './dto/send-mail.dto';
-import { UpdateFileDto } from './dto/upload-file.dto';
+import { UploadFileDto } from './dto/upload-file.dto';
 
 @Controller('admin')
 export class AdminController {
-  constructor(private readonly mailService: MailService, private readonly identService: IdentService) {}
+  constructor(private readonly mailService: MailService, private readonly kycService: KycService) {}
 
   @Post('mail')
   @ApiBearerAuth()
@@ -41,9 +41,9 @@ export class AdminController {
   @UseInterceptors(FilesInterceptor('files'))
   async uploadFile(
     @UploadedFiles() files: Express.Multer.File[],
-    @Body() updateFileDto: UpdateFileDto,
+    @Body() updateFileDto: UploadFileDto,
   ): Promise<boolean> {
-    return this.identService.uploadDocument(updateFileDto.userDataId, files[0], updateFileDto.documentType);
+    return this.kycService.uploadDocument(updateFileDto.userDataId, files[0], updateFileDto.documentType);
   }
 
   @Get('db')
