+++ src/payment/models/sell/sell.service.ts
@@ -8,7 +8,7 @@ import { DepositService } from '../deposit/deposit.service';
 import { User } from '../../../user/models/user/user.entity';
 import { StakingService } from '../staking/staking.service';
 import { Util } from 'src/shared/util';
-import { IdentService } from 'src/user/models/ident/ident.service';
+import { KycService } from 'src/user/models/kyc/kyc.service';
 
 @Injectable()
 export class SellService {
@@ -17,7 +17,7 @@ export class SellService {
     private readonly fiatService: FiatService,
     private readonly depositService: DepositService,
     private readonly stakingService: StakingService,
-    private readonly identService: IdentService,
+    private readonly kycService: KycService,
   ) {}
 
   async getSellByAddress(depositAddress: string): Promise<Sell> {
@@ -35,7 +35,7 @@ export class SellService {
 
   async createSell(userId: number, dto: CreateSellDto): Promise<Sell> {
     // check user data
-    const dataComplete = await this.identService.dataComplete(userId);
+    const dataComplete = await this.kycService.userDataComplete(userId);
     if (!dataComplete) throw new BadRequestException('Ident data incomplete');
 
     // check fiat
