+++ src/user/models/kyc/kyc.controller.ts
@@ -1,0 +1,58 @@
+import { Body, Controller, Get, Post, Query, UploadedFiles, UseGuards, UseInterceptors } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
+import { FilesInterceptor } from '@nestjs/platform-express';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { GetJwt } from 'src/shared/auth/get-jwt.decorator';
+import { JwtPayload } from 'src/shared/auth/jwt-payload.interface';
+import { RoleGuard } from 'src/shared/auth/role.guard';
+import { UserRole } from 'src/shared/auth/user-role.enum';
+import { KycDocument } from 'src/user/services/spider/dto/spider.dto';
+import { LimitRequestDto } from '../limit-request/dto/limit-request.dto';
+import { LimitRequest } from '../limit-request/limit-request.entity';
+import { LimitRequestService } from '../limit-request/limit-request.service';
+import { KycUserDataDto } from '../kyc/dto/kyc-user-data.dto';
+import { KycInfo, KycService } from './kyc.service';
+
+@ApiTags('kyc')
+@Controller('kyc')
+export class KycController {
+  constructor(private readonly kycService: KycService, private readonly limitRequestService: LimitRequestService) {}
+
+  @Get()
+  async getKycProgress(@Query('code') code: string): Promise<KycInfo> {
+    return await this.kycService.getKycStatus(code);
+  }
+
+  @Post()
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
+  async requestKyc(@GetJwt() jwt: JwtPayload): Promise<string> {
+    return await this.kycService.requestKyc(jwt.id).then(JSON.stringify);
+  }
+
+  @Post('data')
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
+  async updateKycData(@GetJwt() jwt: JwtPayload, @Body() data: KycUserDataDto): Promise<boolean> {
+    await this.kycService.updateKycData(jwt.id, data);
+    return true;
+  }
+
+  @Post('limit')
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
+  async increaseLimit(@GetJwt() jwt: JwtPayload, @Body() request: LimitRequestDto): Promise<LimitRequest> {
+    return await this.limitRequestService.increaseLimit(jwt.id, request);
+  }
+
+  @Post('incorporationCertificate')
+  @ApiBearerAuth()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.USER))
+  @UseInterceptors(FilesInterceptor('files'))
+  async uploadIncorporationCertificate(
+    @GetJwt() jwt: JwtPayload,
+    @UploadedFiles() files: Express.Multer.File[],
+  ): Promise<boolean> {
+    return this.kycService.uploadDocument(jwt.id, files[0], KycDocument.INCORPORATION_CERTIFICATE);
+  }
+}
